Rename pmemstream_get_region_runtime
to pmemstream_region_runtime_initialize to make it more clear that
this function might be expensive.
